+++ RELEASE-NOTES.md
@@ -7,8 +7,10 @@ different releases and which versions of PHP and MediaWiki they support, see the
 
 Released on July 4th, 2018.
 
-* Fixed `#display_map` parameter `circles`
-* Fixed `#display_map` parameter `rectangles`
+* Fixed fatal error when using `#display_map` parameter `circles`
+* Fixed fatal error when using `#display_map` parameter `rectangles`
+* Fixed `#display_map` parameter `rectangles` fill color modifier (it is no longer ignored)
+* Fixed `#display_map` parameter `rectangles` fill opacity modifier (it is no longer ignored)
 
 ## Maps 5.5.0